Mv GalaxyContentMeta to models.models
