+++ tests/ansible_galaxy/models/test_content_model.py
@@ -4,7 +4,7 @@ import logging
 import attr
 import pytest
 
-from ansible_galaxy.models import content
+from ansible_galaxy.models.models import GalaxyContentMeta
 from ansible_galaxy.constants import CONTENT_TYPES, CONTENT_TYPE_DIR_MAP
 
 log = logging.getLogger(__name__)
@@ -28,7 +28,7 @@ def test_content_type_dir_map_items():
 
 def test_galaxy_content_meta_no_args():
     try:
-        content.GalaxyContentMeta()
+        GalaxyContentMeta()
     except TypeError as e:
         log.exception(e)
         return
@@ -37,14 +37,14 @@ def test_galaxy_content_meta_no_args():
 
 
 def test_galaxy_content_frozen():
-    content_meta = content.GalaxyContentMeta(namespace='somenamespace',
-                                             name='some_content',
-                                             version='1.0.0',
-                                             src='some_src',
-                                             scm='some_scm',
-                                             content_type='role',
-                                             content_dir='roles',
-                                             path='/dev/null/roles')
+    content_meta = GalaxyContentMeta(namespace='somenamespace',
+                                     name='some_content',
+                                     version='1.0.0',
+                                     src='some_src',
+                                     scm='some_scm',
+                                     content_type='role',
+                                     content_dir='roles',
+                                     path='/dev/null/roles')
     with pytest.raises(attr.exceptions.FrozenInstanceError):
         content_meta.namespace = 'adiffnamespace'
 
@@ -71,34 +71,34 @@ def test_galaxy_content_frozen():
 
 
 def test_galaxy_content_meta():
-    content_meta = content.GalaxyContentMeta(namespace='somenamespace',
-                                             name='some_content',
-                                             version='1.0.0',
-                                             src='some_src',
-                                             scm='some_scm',
-                                             content_type='role',
-                                             content_dir='roles',
-                                             path='/dev/null/roles')
+    content_meta = GalaxyContentMeta(namespace='somenamespace',
+                                     name='some_content',
+                                     version='1.0.0',
+                                     src='some_src',
+                                     scm='some_scm',
+                                     content_type='role',
+                                     content_dir='roles',
+                                     path='/dev/null/roles')
 
     assert content_meta.name == 'some_content'
     assert content_meta.content_type == 'role'
 
 
 def test_galaxy_content_equality():
-    content_meta = content.GalaxyContentMeta(namespace='some_namespace',
-                                             name='some_content',
-                                             version='1.0.0',
-                                             content_type='role')
+    content_meta = GalaxyContentMeta(namespace='some_namespace',
+                                     name='some_content',
+                                     version='1.0.0',
+                                     content_type='role')
 
-    content_meta_newer = content.GalaxyContentMeta(namespace='some_namespace',
-                                                   name='some_content',
-                                                   version='2.3.4',
-                                                   content_type='role')
+    content_meta_newer = GalaxyContentMeta(namespace='some_namespace',
+                                           name='some_content',
+                                           version='2.3.4',
+                                           content_type='role')
 
-    content_meta_newer_dupe = content.GalaxyContentMeta(namespace='some_namespace',
-                                                        name='some_content',
-                                                        version='2.3.4',
-                                                        content_type='role')
+    content_meta_newer_dupe = GalaxyContentMeta(namespace='some_namespace',
+                                                name='some_content',
+                                                version='2.3.4',
+                                                content_type='role')
 
     assert content_meta != content_meta_newer
     assert content_meta_newer != content_meta
